Move XdsClient code out of client_channel tree.
