+++ gRPC-C++.podspec
@@ -259,12 +259,6 @@ Pod::Spec.new do |s|
                       'src/core/ext/filters/client_channel/subchannel.h',
                       'src/core/ext/filters/client_channel/subchannel_interface.h',
                       'src/core/ext/filters/client_channel/subchannel_pool_interface.h',
-                      'src/core/ext/filters/client_channel/xds/xds_api.h',
-                      'src/core/ext/filters/client_channel/xds/xds_bootstrap.h',
-                      'src/core/ext/filters/client_channel/xds/xds_channel.h',
-                      'src/core/ext/filters/client_channel/xds/xds_channel_args.h',
-                      'src/core/ext/filters/client_channel/xds/xds_client.h',
-                      'src/core/ext/filters/client_channel/xds/xds_client_stats.h',
                       'src/core/ext/filters/deadline/deadline_filter.h',
                       'src/core/ext/filters/http/client/http_client_filter.h',
                       'src/core/ext/filters/http/client_authority_filter.h',
@@ -379,6 +373,12 @@ Pod::Spec.new do |s|
                       'src/core/ext/upb-generated/udpa/annotations/versioning.upb.h',
                       'src/core/ext/upb-generated/udpa/data/orca/v1/orca_load_report.upb.h',
                       'src/core/ext/upb-generated/validate/validate.upb.h',
+                      'src/core/ext/xds/xds_api.h',
+                      'src/core/ext/xds/xds_bootstrap.h',
+                      'src/core/ext/xds/xds_channel.h',
+                      'src/core/ext/xds/xds_channel_args.h',
+                      'src/core/ext/xds/xds_client.h',
+                      'src/core/ext/xds/xds_client_stats.h',
                       'src/core/lib/avl/avl.h',
                       'src/core/lib/backoff/backoff.h',
                       'src/core/lib/channel/channel_args.h',
@@ -749,12 +749,6 @@ Pod::Spec.new do |s|
                               'src/core/ext/filters/client_channel/subchannel.h',
                               'src/core/ext/filters/client_channel/subchannel_interface.h',
                               'src/core/ext/filters/client_channel/subchannel_pool_interface.h',
-                              'src/core/ext/filters/client_channel/xds/xds_api.h',
-                              'src/core/ext/filters/client_channel/xds/xds_bootstrap.h',
-                              'src/core/ext/filters/client_channel/xds/xds_channel.h',
-                              'src/core/ext/filters/client_channel/xds/xds_channel_args.h',
-                              'src/core/ext/filters/client_channel/xds/xds_client.h',
-                              'src/core/ext/filters/client_channel/xds/xds_client_stats.h',
                               'src/core/ext/filters/deadline/deadline_filter.h',
                               'src/core/ext/filters/http/client/http_client_filter.h',
                               'src/core/ext/filters/http/client_authority_filter.h',
@@ -869,6 +863,12 @@ Pod::Spec.new do |s|
                               'src/core/ext/upb-generated/udpa/annotations/versioning.upb.h',
                               'src/core/ext/upb-generated/udpa/data/orca/v1/orca_load_report.upb.h',
                               'src/core/ext/upb-generated/validate/validate.upb.h',
+                              'src/core/ext/xds/xds_api.h',
+                              'src/core/ext/xds/xds_bootstrap.h',
+                              'src/core/ext/xds/xds_channel.h',
+                              'src/core/ext/xds/xds_channel_args.h',
+                              'src/core/ext/xds/xds_client.h',
+                              'src/core/ext/xds/xds_client_stats.h',
                               'src/core/lib/avl/avl.h',
                               'src/core/lib/backoff/backoff.h',
                               'src/core/lib/channel/channel_args.h',
