+++ gRPC-Core.podspec
@@ -290,17 +290,6 @@ Pod::Spec.new do |s|
                       'src/core/ext/filters/client_channel/subchannel_interface.h',
                       'src/core/ext/filters/client_channel/subchannel_pool_interface.cc',
                       'src/core/ext/filters/client_channel/subchannel_pool_interface.h',
-                      'src/core/ext/filters/client_channel/xds/xds_api.cc',
-                      'src/core/ext/filters/client_channel/xds/xds_api.h',
-                      'src/core/ext/filters/client_channel/xds/xds_bootstrap.cc',
-                      'src/core/ext/filters/client_channel/xds/xds_bootstrap.h',
-                      'src/core/ext/filters/client_channel/xds/xds_channel.h',
-                      'src/core/ext/filters/client_channel/xds/xds_channel_args.h',
-                      'src/core/ext/filters/client_channel/xds/xds_channel_secure.cc',
-                      'src/core/ext/filters/client_channel/xds/xds_client.cc',
-                      'src/core/ext/filters/client_channel/xds/xds_client.h',
-                      'src/core/ext/filters/client_channel/xds/xds_client_stats.cc',
-                      'src/core/ext/filters/client_channel/xds/xds_client_stats.h',
                       'src/core/ext/filters/client_idle/client_idle_filter.cc',
                       'src/core/ext/filters/deadline/deadline_filter.cc',
                       'src/core/ext/filters/deadline/deadline_filter.h',
@@ -540,6 +529,17 @@ Pod::Spec.new do |s|
                       'src/core/ext/upb-generated/udpa/data/orca/v1/orca_load_report.upb.h',
                       'src/core/ext/upb-generated/validate/validate.upb.c',
                       'src/core/ext/upb-generated/validate/validate.upb.h',
+                      'src/core/ext/xds/xds_api.cc',
+                      'src/core/ext/xds/xds_api.h',
+                      'src/core/ext/xds/xds_bootstrap.cc',
+                      'src/core/ext/xds/xds_bootstrap.h',
+                      'src/core/ext/xds/xds_channel.h',
+                      'src/core/ext/xds/xds_channel_args.h',
+                      'src/core/ext/xds/xds_channel_secure.cc',
+                      'src/core/ext/xds/xds_client.cc',
+                      'src/core/ext/xds/xds_client.h',
+                      'src/core/ext/xds/xds_client_stats.cc',
+                      'src/core/ext/xds/xds_client_stats.h',
                       'src/core/lib/avl/avl.cc',
                       'src/core/lib/avl/avl.h',
                       'src/core/lib/backoff/backoff.cc',
@@ -1154,12 +1154,6 @@ Pod::Spec.new do |s|
                               'src/core/ext/filters/client_channel/subchannel.h',
                               'src/core/ext/filters/client_channel/subchannel_interface.h',
                               'src/core/ext/filters/client_channel/subchannel_pool_interface.h',
-                              'src/core/ext/filters/client_channel/xds/xds_api.h',
-                              'src/core/ext/filters/client_channel/xds/xds_bootstrap.h',
-                              'src/core/ext/filters/client_channel/xds/xds_channel.h',
-                              'src/core/ext/filters/client_channel/xds/xds_channel_args.h',
-                              'src/core/ext/filters/client_channel/xds/xds_client.h',
-                              'src/core/ext/filters/client_channel/xds/xds_client_stats.h',
                               'src/core/ext/filters/deadline/deadline_filter.h',
                               'src/core/ext/filters/http/client/http_client_filter.h',
                               'src/core/ext/filters/http/client_authority_filter.h',
@@ -1274,6 +1268,12 @@ Pod::Spec.new do |s|
                               'src/core/ext/upb-generated/udpa/annotations/versioning.upb.h',
                               'src/core/ext/upb-generated/udpa/data/orca/v1/orca_load_report.upb.h',
                               'src/core/ext/upb-generated/validate/validate.upb.h',
+                              'src/core/ext/xds/xds_api.h',
+                              'src/core/ext/xds/xds_bootstrap.h',
+                              'src/core/ext/xds/xds_channel.h',
+                              'src/core/ext/xds/xds_channel_args.h',
+                              'src/core/ext/xds/xds_client.h',
+                              'src/core/ext/xds/xds_client_stats.h',
                               'src/core/lib/avl/avl.h',
                               'src/core/lib/backoff/backoff.h',
                               'src/core/lib/channel/channel_args.h',
