+++ grpc.gemspec
@@ -209,17 +209,6 @@ Gem::Specification.new do |s|
   s.files += %w( src/core/ext/filters/client_channel/subchannel_interface.h )
   s.files += %w( src/core/ext/filters/client_channel/subchannel_pool_interface.cc )
   s.files += %w( src/core/ext/filters/client_channel/subchannel_pool_interface.h )
-  s.files += %w( src/core/ext/filters/client_channel/xds/xds_api.cc )
-  s.files += %w( src/core/ext/filters/client_channel/xds/xds_api.h )
-  s.files += %w( src/core/ext/filters/client_channel/xds/xds_bootstrap.cc )
-  s.files += %w( src/core/ext/filters/client_channel/xds/xds_bootstrap.h )
-  s.files += %w( src/core/ext/filters/client_channel/xds/xds_channel.h )
-  s.files += %w( src/core/ext/filters/client_channel/xds/xds_channel_args.h )
-  s.files += %w( src/core/ext/filters/client_channel/xds/xds_channel_secure.cc )
-  s.files += %w( src/core/ext/filters/client_channel/xds/xds_client.cc )
-  s.files += %w( src/core/ext/filters/client_channel/xds/xds_client.h )
-  s.files += %w( src/core/ext/filters/client_channel/xds/xds_client_stats.cc )
-  s.files += %w( src/core/ext/filters/client_channel/xds/xds_client_stats.h )
   s.files += %w( src/core/ext/filters/client_idle/client_idle_filter.cc )
   s.files += %w( src/core/ext/filters/deadline/deadline_filter.cc )
   s.files += %w( src/core/ext/filters/deadline/deadline_filter.h )
@@ -459,6 +448,17 @@ Gem::Specification.new do |s|
   s.files += %w( src/core/ext/upb-generated/udpa/data/orca/v1/orca_load_report.upb.h )
   s.files += %w( src/core/ext/upb-generated/validate/validate.upb.c )
   s.files += %w( src/core/ext/upb-generated/validate/validate.upb.h )
+  s.files += %w( src/core/ext/xds/xds_api.cc )
+  s.files += %w( src/core/ext/xds/xds_api.h )
+  s.files += %w( src/core/ext/xds/xds_bootstrap.cc )
+  s.files += %w( src/core/ext/xds/xds_bootstrap.h )
+  s.files += %w( src/core/ext/xds/xds_channel.h )
+  s.files += %w( src/core/ext/xds/xds_channel_args.h )
+  s.files += %w( src/core/ext/xds/xds_channel_secure.cc )
+  s.files += %w( src/core/ext/xds/xds_client.cc )
+  s.files += %w( src/core/ext/xds/xds_client.h )
+  s.files += %w( src/core/ext/xds/xds_client_stats.cc )
+  s.files += %w( src/core/ext/xds/xds_client_stats.h )
   s.files += %w( src/core/lib/avl/avl.cc )
   s.files += %w( src/core/lib/avl/avl.h )
   s.files += %w( src/core/lib/backoff/backoff.cc )
