+++ grpc.gyp
@@ -498,11 +498,6 @@
         'src/core/ext/filters/client_channel/service_config_parser.cc',
         'src/core/ext/filters/client_channel/subchannel.cc',
         'src/core/ext/filters/client_channel/subchannel_pool_interface.cc',
-        'src/core/ext/filters/client_channel/xds/xds_api.cc',
-        'src/core/ext/filters/client_channel/xds/xds_bootstrap.cc',
-        'src/core/ext/filters/client_channel/xds/xds_channel_secure.cc',
-        'src/core/ext/filters/client_channel/xds/xds_client.cc',
-        'src/core/ext/filters/client_channel/xds/xds_client_stats.cc',
         'src/core/ext/filters/client_idle/client_idle_filter.cc',
         'src/core/ext/filters/deadline/deadline_filter.cc',
         'src/core/ext/filters/http/client/http_client_filter.cc',
@@ -628,6 +623,11 @@
         'src/core/ext/upb-generated/udpa/annotations/versioning.upb.c',
         'src/core/ext/upb-generated/udpa/data/orca/v1/orca_load_report.upb.c',
         'src/core/ext/upb-generated/validate/validate.upb.c',
+        'src/core/ext/xds/xds_api.cc',
+        'src/core/ext/xds/xds_bootstrap.cc',
+        'src/core/ext/xds/xds_channel_secure.cc',
+        'src/core/ext/xds/xds_client.cc',
+        'src/core/ext/xds/xds_client_stats.cc',
         'src/core/lib/avl/avl.cc',
         'src/core/lib/backoff/backoff.cc',
         'src/core/lib/channel/channel_args.cc',
@@ -1013,11 +1013,6 @@
         'src/core/ext/filters/client_channel/service_config_parser.cc',
         'src/core/ext/filters/client_channel/subchannel.cc',
         'src/core/ext/filters/client_channel/subchannel_pool_interface.cc',
-        'src/core/ext/filters/client_channel/xds/xds_api.cc',
-        'src/core/ext/filters/client_channel/xds/xds_bootstrap.cc',
-        'src/core/ext/filters/client_channel/xds/xds_channel.cc',
-        'src/core/ext/filters/client_channel/xds/xds_client.cc',
-        'src/core/ext/filters/client_channel/xds/xds_client_stats.cc',
         'src/core/ext/filters/client_idle/client_idle_filter.cc',
         'src/core/ext/filters/deadline/deadline_filter.cc',
         'src/core/ext/filters/http/client/http_client_filter.cc',
@@ -1138,6 +1133,11 @@
         'src/core/ext/upb-generated/udpa/annotations/versioning.upb.c',
         'src/core/ext/upb-generated/udpa/data/orca/v1/orca_load_report.upb.c',
         'src/core/ext/upb-generated/validate/validate.upb.c',
+        'src/core/ext/xds/xds_api.cc',
+        'src/core/ext/xds/xds_bootstrap.cc',
+        'src/core/ext/xds/xds_channel.cc',
+        'src/core/ext/xds/xds_client.cc',
+        'src/core/ext/xds/xds_client_stats.cc',
         'src/core/lib/avl/avl.cc',
         'src/core/lib/backoff/backoff.cc',
         'src/core/lib/channel/channel_args.cc',
