+++ src/core/ext/xds/xds_client_stats.cc
@@ -18,14 +18,14 @@
 
 #include <grpc/support/port_platform.h>
 
-#include "src/core/ext/filters/client_channel/xds/xds_client_stats.h"
+#include "src/core/ext/xds/xds_client_stats.h"
 
 #include <string.h>
 
 #include <grpc/support/atm.h>
 #include <grpc/support/string_util.h>
 
-#include "src/core/ext/filters/client_channel/xds/xds_client.h"
+#include "src/core/ext/xds/xds_client.h"
 
 namespace grpc_core {
