+++ src/python/grpcio/grpc_core_dependencies.py
@@ -75,11 +75,6 @@ CORE_SOURCE_FILES = [
     'src/core/ext/filters/client_channel/service_config_parser.cc',
     'src/core/ext/filters/client_channel/subchannel.cc',
     'src/core/ext/filters/client_channel/subchannel_pool_interface.cc',
-    'src/core/ext/filters/client_channel/xds/xds_api.cc',
-    'src/core/ext/filters/client_channel/xds/xds_bootstrap.cc',
-    'src/core/ext/filters/client_channel/xds/xds_channel_secure.cc',
-    'src/core/ext/filters/client_channel/xds/xds_client.cc',
-    'src/core/ext/filters/client_channel/xds/xds_client_stats.cc',
     'src/core/ext/filters/client_idle/client_idle_filter.cc',
     'src/core/ext/filters/deadline/deadline_filter.cc',
     'src/core/ext/filters/http/client/http_client_filter.cc',
@@ -205,6 +200,11 @@ CORE_SOURCE_FILES = [
     'src/core/ext/upb-generated/udpa/annotations/versioning.upb.c',
     'src/core/ext/upb-generated/udpa/data/orca/v1/orca_load_report.upb.c',
     'src/core/ext/upb-generated/validate/validate.upb.c',
+    'src/core/ext/xds/xds_api.cc',
+    'src/core/ext/xds/xds_bootstrap.cc',
+    'src/core/ext/xds/xds_channel_secure.cc',
+    'src/core/ext/xds/xds_client.cc',
+    'src/core/ext/xds/xds_client_stats.cc',
     'src/core/lib/avl/avl.cc',
     'src/core/lib/backoff/backoff.cc',
     'src/core/lib/channel/channel_args.cc',
