+++ tools/doxygen/Doxyfile.core.internal
@@ -973,17 +973,6 @@ src/core/ext/filters/client_channel/subchannel.h \
 src/core/ext/filters/client_channel/subchannel_interface.h \
 src/core/ext/filters/client_channel/subchannel_pool_interface.cc \
 src/core/ext/filters/client_channel/subchannel_pool_interface.h \
-src/core/ext/filters/client_channel/xds/xds_api.cc \
-src/core/ext/filters/client_channel/xds/xds_api.h \
-src/core/ext/filters/client_channel/xds/xds_bootstrap.cc \
-src/core/ext/filters/client_channel/xds/xds_bootstrap.h \
-src/core/ext/filters/client_channel/xds/xds_channel.h \
-src/core/ext/filters/client_channel/xds/xds_channel_args.h \
-src/core/ext/filters/client_channel/xds/xds_channel_secure.cc \
-src/core/ext/filters/client_channel/xds/xds_client.cc \
-src/core/ext/filters/client_channel/xds/xds_client.h \
-src/core/ext/filters/client_channel/xds/xds_client_stats.cc \
-src/core/ext/filters/client_channel/xds/xds_client_stats.h \
 src/core/ext/filters/client_idle/client_idle_filter.cc \
 src/core/ext/filters/deadline/deadline_filter.cc \
 src/core/ext/filters/deadline/deadline_filter.h \
@@ -1230,6 +1219,17 @@ src/core/ext/upb-generated/udpa/data/orca/v1/orca_load_report.upb.c \
 src/core/ext/upb-generated/udpa/data/orca/v1/orca_load_report.upb.h \
 src/core/ext/upb-generated/validate/validate.upb.c \
 src/core/ext/upb-generated/validate/validate.upb.h \
+src/core/ext/xds/xds_api.cc \
+src/core/ext/xds/xds_api.h \
+src/core/ext/xds/xds_bootstrap.cc \
+src/core/ext/xds/xds_bootstrap.h \
+src/core/ext/xds/xds_channel.h \
+src/core/ext/xds/xds_channel_args.h \
+src/core/ext/xds/xds_channel_secure.cc \
+src/core/ext/xds/xds_client.cc \
+src/core/ext/xds/xds_client.h \
+src/core/ext/xds/xds_client_stats.cc \
+src/core/ext/xds/xds_client_stats.h \
 src/core/lib/README.md \
 src/core/lib/avl/avl.cc \
 src/core/lib/avl/avl.h \
